+++ src/main/java/com/iancmoreno/royaltyparserspring/respository/AsinRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.iancmoreno.royaltyparserspring.model.Asin;
 
 @Repository
-public interface AsinRepository extends JpaRepository<Asin, Integer> {
+public interface AsinRepository extends JpaRepository<Asin, Long> {
 	// enables Spring to create JpaRepository CRUD methods during component scan
 }
