+++ src/main/java/com/iancmoreno/royaltyparserspring/respository/PublisherRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.iancmoreno.royaltyparserspring.model.Publisher;
 
 @Repository
-public interface PublisherRepository extends JpaRepository<Publisher, Integer> {
+public interface PublisherRepository extends JpaRepository<Publisher, Long> {
 	// enables Spring to create JpaRepository CRUD methods during component scan
 }
